+++ authentication_ui_bootstrap5/src/main/resources/static/assets/js/authlib.js
@@ -1,171 +1,354 @@
-/**
- * Redirects to the value set in the hidden input field #contextPath
- */
-function cancel() {
-	location.href = $("meta[name='ctx']").attr("content");
-}
+(function () {
+  // A string of letters, numbers, hyphens, underscores, periods, or plus characters.
+  const usernamePattern = '[A-Za-z0-9.+_\\-]+';
 
-/**
- * Redirects to the user list form after canceling a user edit
- */
-function cancelUserForm() {
-	location.href = $("meta[name='ctx']").attr("content") + "admin/user/list";
-}
+  // A simplified version of the standard email input pattern that excludes exotic email addresses.
+  // See https://developer.mozilla.org/en-US/docs/Web/HTML/Element/input/email#Basic_validation
+  const emailPattern =
+    '[a-zA-Z0-9.+_\\-]+@[a-zA-Z0-9](?:[a-zA-Z0-9\\-]{0,61}[a-zA-Z0-9])?(?:\\.[a-zA-Z0-9](?:[a-zA-Z0-9\\-]{0,61}[a-zA-Z0-9])?)*';
 
-/**
- * Use debounce to wait for keyboard silence before executing a function.
- * For example:
- * 	 $('#username').on('keydown blur change', debounce(myfunction));
- * @see https://john-dugan.com/javascript-debounce/
- * @param func
- * @param wait
- * @param immediate
- * @returns
- */
-function debounce(func, wait, immediate) {
-	var timeout;
-	return function() {
-		var context = this, args = arguments;
-		var later = function() {
-			timeout = null;
-			if (!immediate) func.apply(context, args);
-		};
-		var callNow = immediate && !timeout;
-		clearTimeout(timeout);
-		timeout = setTimeout(later, wait);
-		if (callNow) func.apply(context, args);
-	};
-}
+  // Allows either a bare username or an email address.
+  const mixedPattern = '(?:' + usernamePattern + ')|(?:' + emailPattern + ')';
 
-/**
- * Produces a div with the provided classes.
- * @param text The text to display.
- * @param classNames A space separated list of CSS class names.
- * @returns HTML div with given text and classes.
- */
-function div(text, classNames) {
-	return '<div class="' + classNames + '">' + text + '</div>';
-}
+  /**
+   * Delays a callback until after a noisy event stops firing.
+   * For example:
+   *  someElement.addEventHandler('keydown', debounce(myfunction));
+   * @see https://www.joshwcomeau.com/snippets/javascript/debounce/
+   * @param callback
+   * @param wait
+   * @returns
+   */
+  function debounce(callback, wait) {
+    let timeout = null;
+    return function () {
+      const args = arguments;
+      window.clearTimeout(timeout);
+      timeout = setTimeout(function () {
+        callback.apply(null, args);
+      }, wait);
+    };
+  }
 
-/**
- * Enable or disable save button on save and cancel form fragment.
- */
-function disableSave(booleanValue) {
-	if (booleanValue) {
-		$('.form-controls button[name=save]').attr('disabled', 'disabled');
-	} else {
-		$('.form-controls button[name=save]').attr('disabled', null);
-	}
-}
+  /**
+   * Gets the context path from our standard <meta> tag.
+   */
+  function getContextPath() {
+    const contextPathMeta = document.querySelector('meta[name="ctx"]');
+    let contextPath = '';
 
-$(function() {
-	if (typeof $.fn.DataTable !== 'undefined') {
-		$('.authlib-user-list .users-table').DataTable({
-			columnDefs: [{
-				targets: 0,
-				orderable: false
-			}],
-			order: [[ 1, "asc" ]]
-		});
-	}
+    if (contextPathMeta) {
+      contextPath = contextPathMeta.getAttribute('content');
+    }
 
-	$('.btn.cancel').on('click', cancel);
-	$('.btn.cancelUserForm').on('click', cancelUserForm);
+    if (!contextPath) {
+      console.warn(
+        'Context path was missing or blank. This will probably cause problems.'
+      );
+    }
 
-	if (typeof $.fn.datepicker !== 'undefined') {
-		$("input[name=accountExpirationDate]").datepicker();
-		$("input[name=credentialsExpirationDate]").datepicker();
-	}
+    return contextPath;
+  }
 
-	$('[data-action="popover"]').popover({
-		trigger: "click hover",
-		html: true
-	});
+  /**
+   * Redirects to the root of the application.
+   */
+  function cancelForm() {
+    location.href = getContextPath();
+  }
 
-	var contextPath = $("meta[name='ctx']").attr("content");
+  /**
+   * Redirects to the user list.
+   */
+  function cancelUserForm() {
+    location.href = getContextPath() + 'admin/user/list';
+  }
 
-	/**
-	 * Username type-ahead lookup
-	 * <input id="username" class="lookup-user" />
-	 * Add the `lookup-user` class and after typing a username a lookup
-	 * will be made. A div will be appended with a message about the username.
-	 */
-	$('#username.lookup-user').on('keydown blur change', debounce(function() {
-		var username = $('#username').val();
-		if (username !== null && typeof username !== 'undefined' && username !== '') {
-			// prevent xss - usernames must be lowercase
-			var filteredUsername = username.toLowerCase().replace(/[^a-z0-9]/g, '');
-			// replace username input with filtered username
-			$('#username').val(filteredUsername);
-			$.get(contextPath + 'admin/user/taken/' + encodeURIComponent(filteredUsername), function(json) {
-				$('.username-taken').remove();
-				if (json.taken) {
-					disableSave(true);
-					if ($('.username-taken').length === 0) {
-						$('#username').after(div(filteredUsername + ' is taken', 'username-taken mark bg-danger text-light'));
-					}
-				} else {
-					disableSave(false);
-					if ($('.username-taken').length === 0) {
-						$('#username').after(div(filteredUsername + ' is available', 'username-taken mark bg-success text-light'));
-					}
-				}
-			});
-		} else {
-			$('.username-taken').remove();
-		}
-	}, 500));
+  /**
+   * Sets the pattern used to validate usernames, based on the configured username style.
+   */
+  function setUsernamePattern(usernameInput, usernameStyle) {
+    if (usernameStyle === 'MIXED') {
+      usernameInput.setAttribute('pattern', mixedPattern);
+      usernameInput.dataset.patternErrorMessage = 'Mixed mode error message';
+    } else if (usernameStyle === 'EMAIL') {
+      usernameInput.setAttribute('pattern', emailPattern);
+      usernameInput.dataset.patternErrorMessage = 'Must be an email address';
+    } else {
+      usernameInput.setAttribute('pattern', usernamePattern);
+      usernameInput.dataset.patternErrorMessage =
+        'May only contain the characters a-z0-9.+_-';
+    }
+  }
 
-	// Enable/disable LDAP Lookup based on whether ldapUser exists and is checked
-	var ldapNotChecked = $('#ldap_user').length > 0 && !$('#ldap_user').is(':checked');
-	$('#ldap_lookup').prop("disabled", ldapNotChecked);
+  /**
+   * Returns an appropriate validation error for a username that fails pattern validation.
+   */
+  function getUsernameValidationError(usernameInput) {
+    const characterMessage = 'May only contain the characters a-z0-9.+_-';
+    const emailMessage = 'Not a valid email address';
 
-	$('#ldap_user').on('click', function(e) {
-		$('#ldap_lookup').prop("disabled", !$(this).is(':checked'));
-	});
+    const value = usernameInput.value;
+    const pattern = usernameInput.getAttribute('pattern');
+    if (pattern === mixedPattern) {
+      return value.includes('@') ? emailMessage : characterMessage;
+    } else if (pattern === emailPattern) {
+      return emailMessage;
+    } else {
+      return characterMessage;
+    }
+  }
 
-	// Look up by username in LDAP and prepopulate user fields
-	$('#ldap_lookup').on('click', function(e) {
-		e.preventDefault();
-		var token =  $('input[name="_csrf"]').attr('value');
-	    $.ajaxSetup({
-	        beforeSend: function(xhr) {
-	            xhr.setRequestHeader('X-CSRF-TOKEN', token);
-	        }
-	    });
+  /**
+   * Finds the form group containing the given input element.
+   */
+  function findFormGroupForElement(inputElement) {
+    const formGroupClass = 'form-group';
+    let parentElement = inputElement.parentElement;
+    while (!parentElement.classList.contains(formGroupClass)) {
+      if (parentElement.parentElement) {
+        parentElement = parent.parentElement;
+      }
+    }
 
-		var username = $('#username').val();
-		var obj = {username: username};
+    if (!parentElement.classList.contains(formGroupClass)) {
+      console.error('Element is not enclosed in a form group: ', inputElement);
+      return null;
+    }
 
-		$.post(contextPath + "admin/user/ldapLookup", obj, function(json) {
-			$('.ldap-error').remove();
-			if (json.ldapLookupError) {
-				$('#ldap_lookup').before(div(json.ldapLookupError, 'ldap-error text-danger small'));
-			} else {
-				$('#first_name').val(json.firstName);
-				$('#last_name').val(json.lastName);
-				$('#email').val(json.email);
-				$('#institution').val(json.institution);
-			}
-		});
-	});
+    return parentElement;
+  }
 
-	/**
-	 * Handle displaying inline form validation errors.
-	 * UserController is responsible for sending the BindingResult in the model.
-	 * The mustache template renders hidden input fields with the correct keys
-	 * and messages which are matched with the form element IDs using this JavaScript.
-	 */
-	$('[data-error]').each(function(i, el) {
-		var message = $(el).data('message');
-		$('[name=' + $(el).data('field') + ']')
-			.parent('.form-group')
-			.addClass('has-error')
-			.append('<p class="error-text">' + message + '</p>');
-	})
+  /**
+   * Finds the corresponding `.invalid-feedback` element for the given input.
+   */
+  function findErrorDivForElement(inputElement) {
+    const formGroup = findFormGroupForElement(inputElement);
 
-	/**
-	 * Automatically focus a form element. Used by login and password reset forms.
-	 */
-	$('[data-focus]').first().focus();
-});
+    if (!formGroup) {
+      console.error('Element is not enclosed in a form group: ', inputElement);
+      return;
+    }
+
+    return formGroup.querySelector('.invalid-feedback');
+  }
+
+  /**
+   * Updates the validation feedback for the username input, based on its current validation state.
+   */
+  function updateUsernameFeedback(usernameInput) {
+    const feedbackElement = findErrorDivForElement(usernameInput);
+    if (!feedbackElement) {
+      console.error('No feedback element associated with username input');
+    }
+
+    const validityState = usernameInput.validity;
+    if (!validityState.valid) {
+      const errors = [
+        validityState.customError ? usernameInput.validationMessage : '',
+        validityState.valueMissing ? 'Value must be present' : '',
+        validityState.patternMismatch ? getUsernameValidationError(usernameInput) : '',
+        validityState.tooLong ? 'Username too long (> 50 characters)' : ''
+      ].filter(Boolean);
+      feedbackElement.textContent = errors.join(' ');
+
+      usernameInput.classList.remove('is-valid');
+      usernameInput.classList.add('is-invalid');
+    } else {
+      usernameInput.classList.remove('is-invalid');
+      usernameInput.classList.add('is-valid');
+    }
+  }
+
+  /**
+   * Displays an LDAP search error message.
+   */
+  function showLdapError(message) {
+    const ldapErrorDiv = document.getElementById('ldap_error');
+    if (!ldapErrorDiv) {
+      console.error('LDAP error message element not found');
+      return;
+    }
+
+    ldapErrorDiv.textContent = message;
+    ldapErrorDiv.classList.remove('invisible');
+  }
+
+  /**
+   * Hides the LDAP error message.
+   */
+  function hideLdapError() {
+    const ldapErrorDiv = document.getElementById('ldap_error');
+    if (!ldapErrorDiv) {
+      console.error('LDAP error message element not found');
+      return;
+    }
+
+    ldapErrorDiv.textContent = '';
+    ldapErrorDiv.classList.add('invisible');
+  }
+
+  window.addEventListener('load', function () {
+    //
+    // jQuery plugins
+    //
+
+    // DataTables
+    if (typeof $.fn.DataTable !== 'undefined') {
+      $('.authlib-user-list .users-table').DataTable({
+        columnDefs: [
+          {
+            targets: 0,
+            orderable: false
+          }
+        ],
+        order: [[1, 'asc']]
+      });
+    }
+
+    // Date picker
+    if (typeof $.fn.datepicker !== 'undefined') {
+      $('input[name=accountExpirationDate]').datepicker();
+      $('input[name=credentialsExpirationDate]').datepicker();
+    }
+
+    //
+    // User form event handlers
+    //
+
+    const userForm = document.getElementById('user_form');
+    const dataProperties = userForm ? userForm.dataset : {};
+    const ldapEmailDomain = dataProperties.ldapEmailDomain
+      ? '@' + dataProperties.ldapEmailDomain
+      : null;
+    const usernameStyle = dataProperties.usernameStyle;
+
+    const cancelButtons = document.querySelectorAll('.btn.cancel');
+    for (const cancelButton of cancelButtons) {
+      cancelButton.addEventListener('click', cancelForm);
+    }
+
+    const userFormCancelButtons = document.querySelectorAll('.btn.cancelUserForm');
+    for (const cancelButton of userFormCancelButtons) {
+      cancelButton.addEventListener('click', cancelUserForm);
+    }
+
+    // Check whether the username is already in use
+    const usernameInput = document.getElementById('username');
+    if (usernameInput && usernameInput.classList.contains('lookup-user')) {
+      setUsernamePattern(usernameInput, usernameStyle);
+
+      const searchHandler = debounce(function (_evt) {
+        const searchEndpoint = getContextPath() + 'admin/user/taken/';
+        const username = usernameInput.value.toLowerCase();
+
+        usernameInput.value = username;
+        usernameInput.setCustomValidity('');
+        usernameInput.checkValidity();
+
+        if (!usernameInput.validity.valid) {
+          updateUsernameFeedback(usernameInput);
+          return;
+        }
+
+        if (username) {
+          if (usernameStyle === 'MIXED' && username.endsWith(ldapEmailDomain)) {
+            usernameInput.setCustomValidity(
+              'Username should not be an email for LDAP users.'
+            );
+            updateUsernameFeedback(usernameInput);
+            return;
+          }
+
+          fetch(searchEndpoint + encodeURIComponent(username))
+            .then(response => {
+              if (!response.ok) {
+                const errorMessage = 'Search request failed';
+                usernameInput.setCustomValidity(errorMessage);
+                throw new Error(errorMessage);
+              }
+              return response.json();
+            })
+            .then(jsonData => {
+              if (jsonData.taken) {
+                usernameInput.setCustomValidity('Username is taken');
+              } else {
+                usernameInput.setCustomValidity('');
+              }
+            })
+            .catch(reason => console.error(reason))
+            .finally(() => updateUsernameFeedback(usernameInput));
+        }
+      }, 500);
+
+      usernameInput.addEventListener('input', searchHandler);
+    }
+  });
+
+  // Enable/disable LDAP Lookup based on whether ldapUser exists and is checked
+  const ldapUserCheckbox = document.getElementById('ldap_user');
+  const ldapLookupButton = document.getElementById('ldap_lookup');
+  const ldapNotChecked = ldapUserCheckbox && !ldapUserCheckbox.checked;
+
+  if (ldapLookupButton) {
+    ldapLookupButton.disabled = ldapNotChecked;
+    ldapUserCheckbox.addEventListener('change', function (_evt) {
+      ldapLookupButton.disabled = !this.checked;
+    });
+
+    // Look up by username in LDAP and prepopulate user fields
+    ldapLookupButton.addEventListener('click', function (evt) {
+      evt.preventDefault();
+      hideLdapError();
+      const usernameInput = document.getElementById('username');
+      const username = usernameInput ? usernameInput.value : null;
+
+      if (!username) {
+        return;
+      }
+
+      const csrfTokenInput = document.querySelector('input[name="_csrf"]');
+      const csrfToken = csrfTokenInput ? csrfTokenInput.value : null;
+
+      const ldapLookupEndpoint = getContextPath() + 'admin/user/ldapLookup';
+      const requestBody = new FormData();
+      requestBody.set('username', username);
+
+      fetch(ldapLookupEndpoint, {
+        method: 'post',
+        body: requestBody,
+        headers: {
+          'X-CSRF-TOKEN': csrfToken
+        }
+      })
+        .then(response => {
+          if (!response.ok) {
+            const errorMessage = 'LDAP search request failed';
+            showLdapError(errorMessage);
+            throw new Error(errorMessage);
+          }
+          return response.json();
+        })
+        .then(jsonData => {
+          if (jsonData.ldapLookupError) {
+            showLdapError(jsonData.ldapLookupError);
+          } else {
+            document.getElementById('first_name').value = jsonData.firstName;
+            document.getElementById('last_name').value = jsonData.lastName;
+            document.getElementById('email').value = jsonData.email;
+            document.getElementById('institution').value = jsonData.institution;
+          }
+        })
+        .catch(reason => console.error(reason));
+    });
+  }
+
+  // Display any server validation errors included in hidden elements
+  const serverValidationErrors = document.querySelectorAll('[data-error]');
+  for (const serverError of serverValidationErrors) {
+    const inputElement = document.getElementById(serverError.dataset.field);
+    if (inputElement) {
+      inputElement.classList.add('is-invalid');
+    }
+  }
+})();