+++ authentication_ui_bootstrap5/src/main/resources/static/assets/js/validate.js
@@ -1,7 +1,7 @@
 /**
  * Attaches a listener on any form with the 'needs-validation' class that
  * uses the javascript Constraint Validation API to validate the form.
- * 
+ *
  * source: https://getbootstrap.com/docs/4.0/components/forms/
  */
 
@@ -9,9 +9,10 @@
   'use strict';
   window.addEventListener('load', function() {
     // Fetch all the forms we want to apply custom Bootstrap validation styles to
-    var forms = document.getElementsByClassName('needs-validation');
+    let forms = document.getElementsByClassName('needs-validation');
+
     // Loop over them and prevent submission
-    var validation = Array.prototype.filter.call(forms, function(form) {
+    Array.prototype.filter.call(forms, function(form) {
       form.addEventListener('submit', function(event) {
         if (form.checkValidity() === false) {
           event.preventDefault();